Commit 1 of 4 to bring in the hetero branch to the trunk.  Merged in from:
svn merge -r5440:5448 https://svn.open-mpi.org/svn/ompi/tmp/hetero .

This commit was SVN r5549.

The following SVN revisions from the original message are invalid or
inconsistent and therefore were not cross-referenced:
  r5440
  r5448

+++ src/communicator/comm_publish.c
@@ -114,7 +114,8 @@ int ompi_comm_namepublish ( char *service_name, char *port_name )
     value->tokens[1] = NULL;
     value->num_tokens = 1;
         
-   value->keyvals = (orte_gpr_keyval_t**)malloc(sizeof(orte_gpr_keyval_t *)); 
+
+   value->keyvals = (orte_gpr_keyval_t **)malloc(sizeof(orte_gpr_keyval_t *));
    value->cnt = 1;
 
    value->keyvals[0] = OBJ_NEW(orte_gpr_keyval_t);
@@ -136,7 +137,7 @@ char* ompi_comm_namelookup ( char *service_name )
     char *token[2], *key[2];
     orte_gpr_keyval_t **keyvals=NULL;
     orte_gpr_value_t **values;
-    int cnt=0;
+    size_t cnt=0;
     char *stmp=NULL;
     int ret;
